requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
attrs==22.2.0; python_version >= '3.6'
certifi==2022.12.7; python_version >= '3.6'
charset-normalizer==3.1.0; python_full_version >= '3.7.0'
contourpy==1.0.7; python_version >= '3.8'
cycler==0.11.0; python_version >= '3.6'
exceptiongroup==1.1.0; python_version < '3.11'
fonttools==4.39.0; python_version >= '3.8'
idna==3.4; python_version >= '3.5'
iniconfig==2.0.0; python_full_version >= '3.7.0'
kiwisolver==1.4.4; python_full_version >= '3.7.0'
matplotlib==3.7.1
numpy==1.24.2
packaging==23.0; python_full_version >= '3.7.0'
pandas==1.5.3
pillow==9.4.0; python_full_version >= '3.7.0'
pluggy==1.0.0; python_version >= '3.6'
pyparsing==3.0.9; python_full_version >= '3.6.8'
pytest==7.2.2
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2022.7.1
pyyaml==6.0
quantlib==1.29
repackage==0.7.3
requests==2.28.2
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tomli==2.0.1; python_version < '3.11'
urllib3==1.26.14; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
